rename coll view
