+++ packages/editor-ui/src/views/TemplatesCollectionView.vue
